requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
cachelib==0.6.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
git+https://github.com/George-Sava/flask-jwt#egg=Flask-JWT
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
git+https://github.com/George-Sava/pyjwt#egg=PyJWT
git+https://github.com/marshmallow-code/flask-marshmallow#egg=flask-marshmallow
git+https://github.com/glic3rinu/passlib#egg=passlib
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
typing_extensions==4.0.1
Werkzeug==2.0.2
uwsgi
